refactor: split the openlayers module by concern; native GeoJSON upload dialog
The 660-line openlayers/index.ts mixed editing tools, popups, view
handling, basemap selection, geolocation and z-coordinate helpers in
one module. Split it into focused modules (edit, popover, view,
basemap, geolocation, zcoords, upload-dialog) behind the existing
barrel, so importers are unchanged.

The GeoJSON upload modal moves from jQuery UI .dialog() to the native
<dialog> element with a method="dialog" form, removing the plugin's
last jQuery UI dependency. The dialog buttons are now server-rendered
from the existing gtt_js.modal locale keys, and the .ui-widget /
.ui-dialog CSS is replaced by .gtt-dialog styles with a ::backdrop.

Also:
- file uploads now fill the textarea with the raw GeoJSON text
  instead of a JSON.stringify()-quoted string
- invalid GeoJSON shows a map notification instead of throwing
- add the missing gtt_js.control.edit_mode locale key (the edit
  toggle's tooltip rendered as "undefined")
- drop dead code: the unused ExtendedTooltip.prevHTML interface and
  an empty geolocation change handler
