+++ src/styles/scss/app.scss
@@ -237,22 +237,34 @@ i[id^='icon_settings_tracker_'] {
   }
 }
 
-.ui-widget.ui-widget-content {
-  max-width: 510px;
-  padding: 0.5em;
-}
+/** GeoJSON upload dialog (native <dialog>) */
+.gtt-dialog {
+  width: 380px;
+  max-width: 90vw;
+  padding: 1em;
+  border: 1px solid #ccc;
+  border-radius: 4px;
 
-.ui-dialog .ui-dialog-buttonpane button {
-  height: inherit;
-}
+  &::backdrop {
+    background: rgba(0, 0, 0, 0.4);
+  }
 
-.ui-dialog .ui-dialog-content {
-  padding: 0.5em 0 0 0;
-}
+  h3.title {
+    margin-top: 0;
+  }
 
-.ui-dialog .ui-dialog-content textarea {
-  width: 100%;
-  height: 300px;
+  textarea {
+    display: block;
+    width: 100%;
+    height: 300px;
+    box-sizing: border-box;
+    margin-bottom: 0.5em;
+  }
+
+  p.buttons {
+    margin-bottom: 0;
+    text-align: right;
+  }
 }
 
 /**
